disk usage fix

+++ internal/services/mcp/fs_disk_usage_test_bsize_linux_32bit.go
@@ -2,12 +2,12 @@
 // Licensed under the Apache License, Version 2.0 (the "License").
 // See LICENSE for full text.
 
-//go:build linux && (amd64 || arm64 || ppc64 || ppc64le || mips64 || mips64le || s390x || loong64)
+//go:build linux && !(amd64 || arm64 || ppc64 || ppc64le || mips64 || mips64le || s390x || loong64)
 
 package mcp
 
 import "syscall"
 
 func setStatfsBsize(stat *syscall.Statfs_t, bsize uint32) {
-	stat.Bsize = int64(bsize)
+	stat.Bsize = int32(bsize)
 }
